blog/requirements.txt
django
django-ckeditor
Pillow
django-crispy-forms
crispy-bootstrap5